LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-during-abort-event-expected.txt


FAIL XMLHttpRequest: open() during abort event - abort() called from upload.onloadstart assert_array_equals: property 5, expected "readyState after abort() 1" but got "readyState after abort() 0"
